Feature/sound relay config (#24)
* Update control panel so can configure with sound relay index

* Partial implementation of sounds within BoatElectronics

* Implement horn (fog etc) sounds etc including SoS, narrow channel, maneuvering

* Updates from AI code review

+++ BoatControlPanel/WarningCommandHandler.cpp
@@ -28,7 +28,7 @@ bool WarningCommandHandler::handleCommand(SerialCommandManager* sender, const St
 
     if (cmd == WarningsActive && paramCount == 0)
     {
-        int count = 0;
+        uint8_t count = 0;
         for (uint8_t i = 1; i < WarningCount; i++)
         {
             WarningType type = static_cast<WarningType>(i);
